+++ src/pamela/cli.clj
@@ -21,7 +21,9 @@
             [me.raynes.fs :as fs]
             [pamela.utils :refer [repl? set-cwd!
                                   input-file output-file]]
+            [plan-schema.utils :refer [fs-get-path]]
             [pamela.parser :as parser]
+            [pamela.unparser :as unparser]
             [pamela.htn :as htn]
             [pamela.tpn :as tpn]
             [clojure.tools.logging :as log]
@@ -67,7 +69,7 @@
   "Load model(s) and build intermediate representation (IR)"
   {:added "0.3.0"}
   [options]
-  (let [{:keys [input output]} options
+  (let [{:keys [input output source]} options
         ir (parser/parse options)]
     (if (:error ir)
       (do
@@ -75,6 +77,23 @@
         1)
       (do
         (output-file output "edn" ir)
+        (when source
+          (output-file source "raw" (unparser/unparse ir)))
+        0))))
+
+(defn unparse-model
+  "Load model(s) and build intermediate representation (IR)"
+  {:added "0.3.0"}
+  [options]
+  (let [{:keys [input output source]} options
+        ir (read-string (slurp (fs-get-path (first input))))]
+    ;; assumes there is exactly one input file
+    (if source
+      (do
+        (log/errorf "specifiy --output instead of --source for the unparse action")
+        1)
+      (do
+        (output-file output "raw" (unparser/unparse ir))
         0))))
 
 (defn parse-model
@@ -132,6 +151,7 @@
   "Valid PAMELA command line actions"
   {"check" (var check-model)
    "build" (var build-model)
+   "unparse" (var unparse-model)
    "tpn" (var tpn)
    "htn" (var htn)})
 
@@ -147,17 +167,21 @@
 (def #^{:added "0.2.0"}
   cli-options
   "Command line options"
-  [["-h" "--help" "Print usage"]
-   ["-V" "--version" "Print PAMELA version"]
-   ["-v" "--verbose" "Increase verbosity"
-    :default 0
-    :assoc-fn (fn [m k _] (update-in m [k] inc))]
+  [["-V" "--version" "Print PAMELA version"]
+   ["-a" "--magic MAGIC" "Magic lvar initializtions"
+    :default nil
+    :parse-fn #(input-file %)
+    :validate [#(or (nil? %) (fs/exists? %))
+               "MAGIC file does not exist"]]
+   ["-b" "--output-magic OUTPUT-MAGIC" "Output magic file"
+    :default nil]
    ["-c" "--construct-tpn CFM " "Construct TPN using class C field F method M (as C:F:M)"]
    ["-f" "--file-format FORMAT" "Output file format [json]"
     :default "json"
     :validate [#(contains? output-formats %)
                (str "FORMAT not supported, must be one of "
                  (vec output-formats))]]
+   ["-h" "--help" "Print usage"]
    ["-i" "--input INPUT" "Input file(s) (or - for STDIN)"
     :default ["-"]
     :parse-fn #(input-file %)
@@ -176,14 +200,11 @@
                                        log-levels)))))]
    ["-o" "--output OUTPUT" "Output file (or - for STDOUT)"
     :default "-"]
-   ["-a" "--magic MAGIC" "Magic lvar initializtions"
-    :default nil
-    :parse-fn #(input-file %)
-    :validate [#(or (nil? %) (fs/exists? %))
-               "MAGIC file does not exist"]]
-   ["-b" "--output-magic OUTPUT-MAGIC" "Output magic file"
-    :default nil]
+   ["-s" "--source PAMELA" "source as reconstructed from IR"]
    ["-t" "--root-task ROOTTASK" "Label for HTN root-task [main]"]
+   ["-v" "--verbose" "Increase verbosity"
+    :default 0
+    :assoc-fn (fn [m k _] (update-in m [k] inc))]
    ])
 
 (defn usage
@@ -244,7 +265,7 @@
         (parse-opts args cli-options)
         {:keys [help version verbose construct-tpn
                 file-format input log-level output root-task
-                magic output-magic]} options
+                source magic output-magic]} options
         log-level (keyword (or log-level "warn"))
         _ (plog/initialize log-level (apply pr-str args))
         cmd (first arguments)
@@ -277,6 +298,7 @@
       (println "file-format:" file-format)
       (println "input:" input)
       (println "output:" output)
+      (println "source:" source)
       (println "magic:" magic)
       (println "output-magic:" output-magic)
       (println "root-task:" root-task)